requirements.txt
streamlit>=1.30
Pillow>=10.0